cleanup process and move wrapper functions to respective core modules
