+++ pyrate/core/covariance.py
@@ -19,7 +19,6 @@ Variance/Covariance matrix functionality.
 """
 # coding: utf-8
 from os.path import basename, join
-import logging
 from numpy import array, where, isnan, real, imag, sqrt, meshgrid
 from numpy import zeros, vstack, ceil, mean, exp, reshape
 from numpy.linalg import norm
@@ -27,13 +26,15 @@ import numpy as np
 from scipy.fftpack import fft2, ifft2, fftshift
 from scipy.optimize import fmin
 
-from pyrate.core import shared, ifgconstants as ifc, config as cf
-from pyrate.core.shared import PrereadIfg
+from pyrate.core import shared, ifgconstants as ifc, config as cf, mpiops
+from pyrate.core.shared import PrereadIfg, Ifg
 from pyrate.core.algorithm import master_slave_ids
 from pyrate.core.logger import pyratelogger as log
 
 # pylint: disable=too-many-arguments
 # distance division factor of 1000 converts to km and is needed to match legacy output
+
+MASTER_PROCESS = 0
 DISTFACT = 1000
 
 
@@ -326,3 +327,48 @@ def get_vcmt(ifgs, maxvar):
     std = sqrt(maxvar).reshape((nifgs, 1))
     vcm_t = std * std.transpose()
     return vcm_t * vcm_pat
+
+
+def maxvar_vcm_calc_wrapper(params):
+    """
+    MPI wrapper for maxvar and vcmt computation
+    """
+    preread_ifgs = params[cf.PREREAD_IFGS]
+    ifg_paths = [ifg_path.sampled_path for ifg_path in params[cf.INTERFEROGRAM_FILES]]
+    log.info('Calculating the temporal variance-covariance matrix')
+    process_indices = mpiops.array_split(range(len(ifg_paths)))
+
+    def _get_r_dist(ifg_path):
+        """
+        Get RDIst class object
+        """
+        ifg = Ifg(ifg_path)
+        ifg.open()
+        r_dist = RDist(ifg)()
+        ifg.close()
+        return r_dist
+
+    r_dist = mpiops.run_once(_get_r_dist, ifg_paths[0])
+    prcs_ifgs = mpiops.array_split(ifg_paths)
+    process_maxvar = []
+    for n, i in enumerate(prcs_ifgs):
+        log.debug('Calculating maxvar for {} of process ifgs {} of total {}'.format(n+1, len(prcs_ifgs), len(ifg_paths)))
+        process_maxvar.append(cvd(i, params, r_dist, calc_alpha=True, write_vals=True, save_acg=True)[0])
+    if mpiops.rank == MASTER_PROCESS:
+        maxvar = np.empty(len(ifg_paths), dtype=np.float64)
+        maxvar[process_indices] = process_maxvar
+        for i in range(1, mpiops.size):  # pragma: no cover
+            rank_indices = mpiops.array_split(range(len(ifg_paths)), i)
+            this_process_ref_phs = np.empty(len(rank_indices), dtype=np.float64)
+            mpiops.comm.Recv(this_process_ref_phs, source=i, tag=i)
+            maxvar[rank_indices] = this_process_ref_phs
+    else:  # pragma: no cover
+        maxvar = np.empty(len(ifg_paths), dtype=np.float64)
+        mpiops.comm.Send(np.array(process_maxvar, dtype=np.float64), dest=MASTER_PROCESS, tag=mpiops.rank)
+
+    mpiops.comm.barrier()
+    maxvar = mpiops.comm.bcast(maxvar, root=0)
+    vcmt = mpiops.run_once(get_vcmt, preread_ifgs, maxvar)
+    log.debug("Finished maxvar and vcm calc!")
+    params[cf.MAXVAR], params[cf.VCMT] = maxvar, vcmt
+    return maxvar, vcmt
