+++ pyrate/core/mst.py
@@ -19,6 +19,8 @@ This Python module implements the minimum spanning tree
 functionality for selecting interferometric observations.
 """
 # pylint: disable=invalid-name
+from os.path import join
+
 from itertools import product
 from numpy import array, nan, isnan, float32, empty, sum as nsum
 import numpy as np
@@ -28,7 +30,7 @@ from joblib import Parallel, delayed
 
 from pyrate.core.algorithm import ifg_date_lookup
 from pyrate.core.algorithm import ifg_date_index_lookup
-from pyrate.core import config as cf
+from pyrate.core import config as cf, mpiops
 from pyrate.core.shared import IfgPart, create_tiles
 from pyrate.core.shared import joblib_log_level
 from pyrate.core.logger import pyratelogger as log
@@ -273,3 +275,28 @@ def _minimum_spanning_edges_from_mst(edges):
     T = nx.minimum_spanning_tree(g_nx)  # step ifglist_mst in make_mstmat.m
     edges = T.edges()
     return edges, g_nx
+
+
+def mst_calc_wrapper(params):
+    """
+    MPI wrapper function for MST calculation
+    """
+    tiles = params[cf.TILES]
+    preread_ifgs = params[cf.PREREAD_IFGS]
+    dest_tifs = [ifg_path.sampled_path for ifg_path in params[cf.INTERFEROGRAM_FILES]]
+    process_tiles = mpiops.array_split(tiles)
+    log.info('Calculating minimum spanning tree matrix')
+
+    def _save_mst_tile(tile, i, preread_ifgs):
+        """
+        Convenient inner loop for mst tile saving
+        """
+        mst_tile = mst_multiprocessing(tile, dest_tifs, preread_ifgs, params)
+        # locally save the mst_mat
+        mst_file_process_n = join(params[cf.TMPDIR], 'mst_mat_{}.npy'.format(i))
+        np.save(file=mst_file_process_n, arr=mst_tile)
+
+    for t in process_tiles:
+        _save_mst_tile(t, t.index, preread_ifgs)
+    log.debug('Finished mst calculation for process {}'.format(mpiops.rank))
+    mpiops.comm.barrier()
