+++ tests/test_mpi.py
@@ -22,6 +22,11 @@ import shutil
 import numpy as np
 import os
 from pathlib import Path
+
+import pyrate.core.covariance
+import pyrate.core.orbital
+import pyrate.core.ref_phs_est
+import pyrate.core.refpixel
 from pyrate import process, prepifg, conv2tif, configuration
 from pyrate.core import mpiops, config as cf
 from tests import common
@@ -50,11 +55,11 @@ def test_vcm_legacy_vs_mpi(mpisync, tempdir, roipac_or_gamma_conf):
     params[cf.INTERFEROGRAM_FILES].pop()
     process._update_params_with_tiles(params)
     process._create_ifg_dict(params=params)
-    process._ref_pixel_calc(params)
-    process._orb_fit_calc(params)
-    process._ref_phase_est_wrapper(params)
+    pyrate.core.refpixel.ref_pixel_calc_wrapper(params)
+    pyrate.core.orbital.orb_fit_calc_wrapper(params)
+    pyrate.core.ref_phs_est.ref_phase_est_wrapper(params)
 
-    maxvar, vcmt = process._maxvar_vcm_calc(params)
+    maxvar, vcmt = pyrate.core.covariance.maxvar_vcm_calc_wrapper(params)
 
     # phase data after ref pixel has changed due to commit bf2f7ebd
     # Legacy tests won't match anymore
