+++ tests/test_stackrate.py
@@ -28,6 +28,8 @@ import numpy as np
 from numpy.testing import assert_array_almost_equal, assert_array_equal
 
 import pyrate.core.orbital
+import pyrate.core.ref_phs_est
+import pyrate.core.refpixel
 import tests.common
 from pyrate.core import shared, config as cf, covariance as vcm_module
 from pyrate.core.stack import stack_rate_pixel, mask_rate
@@ -129,7 +131,7 @@ class LegacyEqualityTest(unittest.TestCase):
         ifgs = pre_prepare_ifgs(dest_paths, params)
         mst_grid = tests.common.mst_calculation(dest_paths, params)
 
-        refx, refy = process._ref_pixel_calc(params)
+        refx, refy = pyrate.core.refpixel.ref_pixel_calc_wrapper(params)
         params[cf.REFX] = refx
         params[cf.REFY] = refy
 
@@ -139,7 +141,7 @@ class LegacyEqualityTest(unittest.TestCase):
         for ifg in ifgs:
             ifg.close()
         process._update_params_with_tiles(params)
-        _, ifgs = process._ref_phase_est_wrapper(params)
+        _, ifgs = pyrate.core.ref_phs_est.ref_phase_est_wrapper(params)
         ifgs[0].open()
         r_dist = vcm_module.RDist(ifgs[0])()
         ifgs[0].close()
